In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,16 +5,16 @@
  */
 int main(void)
 {
-	int n1 = 1;
-	int n2 = 2;
-	int n3;
+	unsigned long n1 = 1;
+	unsigned long n2 = 2;
+	unsigned long n3;
 	int i;
 
-	printf("%d, %d", n1, n2);
+	printf("%lu, %lu", n1, n2);
 	for (i = 0; i < 48; i++)
 	{
 		n3 = n1 + n2;
-		printf(", %d", n3);
+		printf(", %lu", n3);
 		n1 = n2;
 		n2 = n3;
 	}
